refactor(lib): moved lib/server code to server/lib folder
