+++ src/server/services/problems.ts
@@ -2,7 +2,7 @@ import { db } from "@/db";
 import { classes, problems, users, users_to_classes } from "@/db/schema";
 import { ClassSchema, ProblemSchema } from "@/db/validation";
 import { UserColumns } from "@/db/validation/schemas/user";
-import { pagination } from "@/lib/server/pagination";
+import { pagination } from "@/server/lib/pagination";
 import { TRPCError } from "@trpc/server";
 import { and, desc, eq, getTableColumns, sql } from "drizzle-orm";
 import { z } from "zod/v4";
